Fixing the nextj ssr error
